refactor applicant tests into their own package under the commands package and parser package
