+++ src/test/java/seedu/address/logic/commands/applicant/EditApplicantCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.applicant;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -18,6 +18,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.applicant.EditApplicantCommand;
 import seedu.address.logic.commands.applicant.EditApplicantCommand.EditApplicantDescriptor;
 import seedu.address.model.AddressBook;
